requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docstring_parser==0.16
Flask==3.0.3
flask-sock==0.7.0
Flask-Sockets==0.2.1
frozenlist==1.4.1
gevent==24.2.1
gevent-websocket==0.10.1
google-ai-generativelanguage==0.6.1
google-api-core==2.0.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.47.0
google-cloud-bigquery==3.20.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-speech==2.0.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain-core==0.1.42
langchain-google-genai==1.0.2
langchain-google-vertexai==0.1.3
langsmith==0.1.42
libcst==1.3.1
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
orjson==3.10.0
packaging==23.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
shapely==2.0.3
simple-websocket==1.0.0
six==1.16.0
tenacity==8.2.3
tqdm==4.66.2
twilio==9.0.4
types-protobuf==4.25.0.20240410
types-requests==2.31.0.20240406
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
websocket==0.2.1
Werkzeug==3.0.2
wsproto==1.2.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.2
